All Modules return now the actuall object.
For example `require('Fx/Fx')` would return the Fx class
or 'Fx/Fx.Tween' would return the Tween class.

'Core/Core' returns an object with all the functions and objects.

Core implements a simple AMD loader that is necessary to load all the modules the old way or without an external loader
The compatibility block `<!amd>` is used to export all modules in the global scope like it used to be.