Also extend the global object if there is an AMD loader with a thruthy define.amd.
This is especially important for 3rd party MooTools plugins.
In case of using a loader that defers execution of the module factory, it is very important that these modules are required first, before loading other scripts that might depend on the global objects.

When using a full built version, just using one <script> tag, and requiring all defined modules would do this, once this is done, other <script> tags might be loaded.
When loading all files independently, some AMD loaders (requirejs) have an order! plugin, that makes sure the files are loaded in order.

Puting stuff in the global object must be inside the factory function, because outside this function, there is no 'standardized' require() function we can rely on.